refactor recombine to use: fetch + mend
